requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-qdrant>=0.2.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
qdrant-client>=1.12.0
pypdf>=5.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
colorama>=0.4.6
